This branch will refactor the /locations routes from app.py into a blueprint file called locations.py. Some of the local classes and variables had to be moved from app.py to their own separate files, which are env.py, model.py, and schema.py. The /historian routes were also moved to historian.py, but I was unable to test them due to errors that occurred with the /historian functionality in the main branch (specifically Value Errors). The _get_historical_data_query module also had circular import issues when imported at the top of the file in historian.py, so I had to import the module in each of the routes. It is unclear if this solution works properly.
