+++ Lumen-Server/env.py
@@ -1,0 +1,2 @@
+mode = ''
+config = {}
